pyproject.toml
[tool.poetry]
name = "mychatgpt-demo"
version = "0.1.0"
description = ""
authors = ["Jared Zoneraich <jszoneraich@gmail.com>"]
readme = "README.md"
packages = [{include = "mychatgpt_demo"}]

[tool.poetry.dependencies]
python = "^3.9"
promptlayer = "^0.5.0"
openai = "^1.14.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
